Refactor OpenAI API calls. Add anthropic API support. Unified model multiplier integral deduction. Support API proxy function. Modify the login page style.

+++ server/src/routes/admin/token.ts
@@ -2,8 +2,8 @@ import {Router} from "express";
 import utils from "../../utils";
 import ApiResponse from "../../utils/response";
 import {Token} from "../../models/Token";
-import {addUsageCheckTask, getKeyUsage} from "../../chatgpt/key_usage";
-import {putClient, removeClient} from "../../chatgpt";
+import {addUsageCheckTask, getKeyUsage} from "../../ai/openai/key_usage";
+import {supplierClientAgent} from "../../ai";
 
 const router = Router();
 router.get('/', async function (req, res) {
@@ -21,45 +21,56 @@ router.delete('/:id', async function (req, res) {
     res.json(ApiResponse.miss());
     return;
   }
+  let token = await Token.findByPk(id);
+  if (!token)
+    return;
+  supplierClientAgent.removeClient(token.toJSON());
   res.json(ApiResponse.success(await Token.destroy({
     where: {
       id
     }
-  }).then(() => {
-    removeClient(Number(id));
   })));
 });
 
 router.post('/', async function (req, res) {
-  const {key, host, remarks, models, status} = req.body;
+  const {key, host, remarks, models, status, supplier = 'openai'} = req.body;
   if (!key || !host || !models) {
     res.json(ApiResponse.miss());
     return;
   }
-  res.json(ApiResponse.success(await Token.add({key, host, remarks, models, status}).then((token) => {
-    putClient(token.toJSON());
+  res.json(ApiResponse.success(await Token.add({key, host, remarks, models, status, supplier}).then((token) => {
+    supplierClientAgent.putClient(token.toJSON());
     return token;
   })));
 });
 
 router.put('/', async function (req, res) {
-  const {id, key, host, remarks, models, status} = req.body;
+  const {id, key, host, remarks, models, status, supplier = 'openai'} = req.body;
   if (!id || !key || !host || !models) {
     res.json(ApiResponse.miss());
     return;
   }
-  res.json(ApiResponse.success(await Token.edit({id, key, host, remarks, models, status}).then((data) => {
-    putClient(data?.[0]?.toJSON());
-    return data?.[0];
-  })))});
+  res.json(ApiResponse.success(await Token.edit({id, key, host, remarks, models, status, supplier})
+    .then((data) => {
+      if (status === 0) {
+        supplierClientAgent.removeClient(data?.[0]?.toJSON());
+      } else {
+        supplierClientAgent.putClient(data?.[0]?.toJSON());
+      }
+      return data?.[0];
+    }))
+  )
+});
 router.post('/check', async function (req, res) {
   const {key, host, all} = req.body;
   if (all) {
     const tokens = await Token.findAll({
       limit: 100,
       where: {
-        status: 1
-      }
+        status: 1,
+        supplier: 'openai'
+      },
+      raw: true
     });
     tokens.forEach((token) => {
       addUsageCheckTask({
